bbk_miniapps/backend/requirements.txt
fastapi
uvicorn
python-dotenv
gspread
oauth2client
aiogram
uvloop
httptools